+++ Configuration/TypoScript/GridElements/constants.ts
@@ -1,3 +1,0 @@
-# === Includes ===
-
-<INCLUDE_TYPOSCRIPT: source="FILE: ./NoteGridElement/constants.ts">