+++ Configuration/TypoScript/Library/T3v/setup.ts
@@ -1,3 +1,0 @@
-# === Includes ===
-
-<INCLUDE_TYPOSCRIPT: source="FILE: ./Cookies/setup.ts">